refactor:move loadFile.service into service folder
#2294
