+++ visualization/app/codeCharta/services/LoadFile/loadFile.service.ts
@@ -1,16 +1,16 @@
 import { Injectable, Inject } from "@angular/core"
-import { CCFileValidationResult, checkErrors, checkWarnings } from "./util/fileValidator"
-import { NodeDecorator } from "./util/nodeDecorator"
-import { setFiles, setStandardByNames } from "./state/store/files/files.actions"
-import { FileSelectionState, FileState } from "./model/files/files"
-import { getCCFile } from "./util/fileHelper"
-import { NameDataPair } from "./codeCharta.model"
-import { referenceFileSelector } from "./state/selectors/referenceFile/referenceFile.selector"
+import { CCFileValidationResult, checkErrors, checkWarnings } from "../../util/fileValidator"
+import { NodeDecorator } from "../../util/nodeDecorator"
+import { setFiles, setStandardByNames } from "../../state/store/files/files.actions"
+import { FileSelectionState, FileState } from "../../model/files/files"
+import { getCCFile } from "../../util/fileHelper"
+import { NameDataPair } from "../../codeCharta.model"
+import { referenceFileSelector } from "../../state/selectors/referenceFile/referenceFile.selector"
 import { MatDialog } from "@angular/material/dialog"
-import { ErrorDialogComponent } from "./ui/dialogs/errorDialog/errorDialog.component"
-import { loadFilesValidationToErrorDialog } from "./util/loadFilesValidationToErrorDialog"
-import { Store } from "./state/angular-redux/store"
-import { State } from "./state/angular-redux/state"
+import { ErrorDialogComponent } from "../../ui/dialogs/errorDialog/errorDialog.component"
+import { loadFilesValidationToErrorDialog } from "../../util/loadFilesValidationToErrorDialog"
+import { Store } from "../../state/angular-redux/store"
+import { State } from "../../state/angular-redux/state"
 import { tap } from "rxjs"
 
 @Injectable({ providedIn: "root" })
