+++ visualization/app/codeCharta/util/nodePathHelper.ts
@@ -1,4 +1,4 @@
-import { LoadFileService } from "../loadFile.service"
+import { LoadFileService } from "../services/LoadFile/loadFile.service"
 
 export function getUpdatedBlacklistItemPath(fileName: string, path: string) {
 	if (isAbsoluteRootPath(path)) {
